FF-357 #comment More attempts to get the linking working
