+++ src/encoded/schemas/sop_map.json
@@ -4,6 +4,7 @@
     "id": "/profiles/sop_map.json",
     "$schema": "http://json-schema.org/draft-04/schema#",
     "type": "object",
+    "required": ["associated_item_type"],
     "additionalProperties": false,
     "identifyingProperties": ["uuid", "aliases", "mapid"],
     "mixinProperties": [
@@ -45,6 +46,11 @@
                 "description": "Identifying field value",
                 "type": "string"
             }
+        },
+        "mapid": {
+            "title": "Map ID",
+            "description": "An autogenerated unique ID to allow mapping of Items to an SOP map",
+            "type": "string"
         },
         "description": {
             "title": "Description",
