+++ src/encoded/types/__init__.py
@@ -268,35 +268,6 @@ class WorkflowMapping(Item):
 
 
 @collection(
-    name='sop_maps',
-    unique_key='mapid',
-    properties={
-        'title': 'SOP and field mappings',
-        'description': 'Listing of SOPs with the default values for fields from them',
-    })
-class SopMap(Item):
-    """The SopFields class that lists the default values of fields if the SOP is followed."""
-
-    item_type = 'sop_map'
-    name_key = 'mapid'
-    schema = load_schema('encoded:schemas/sop_map.json')
-
-    @calculated_property(schema={
-        "title": "ID",
-        "description": "Composite uniquely identifying ID for sop map",
-        "type": "string"
-    })
-    def mapid(self, request, schema_version, associated_item_type, id_values=None):
-        delim = '_'
-        id_string = associated_item_type + delim
-        if id_values is not None:
-            for f in id_values:
-                f = ''.join(f.split())
-                id_string = id_string + f + delim
-        return id_string + schema_version
-
-
-@collection(
     name='targets',
     properties={
         'title': 'Targets',
